+++ y/y/core/Chrono.h
@@ -31,7 +31,7 @@ namespace core {
 class Duration {
 	static constexpr Duration div(double s, double div) {
 		double nano_div = 1000000000 / div;
-		u64 secs = s / div;
+		u64 secs = u64(s / div);
 		s -= secs * div;
 		return Duration(secs, u32(s * nano_div));
 	}
